+++ models/staging/reference_data/stg_reference_data__product_catalog.sql
@@ -10,8 +10,8 @@ renamed as (
         cast(product_type as string)      as product_type,
         cast(product_name as string)      as product_name,
         cast(product_category as string)  as product_category,
-        cast(max_ltv as numeric)           as max_ltv,
-        cast(min_interest_rate as numeric) as min_interest_rate,
+        cast(max_ltv as double)           as max_ltv,
+        cast(min_interest_rate as double) as min_interest_rate,
         cast(max_term_months as integer)   as max_term_months
 
     from source
